DBENGINE v2 - improvements 2 (#14257)
* allow extents to be merged for as long as possible

* do not block the event loop while recalculating retention due to datafile rotation

* buffers are incrementally cleaned up, every second, by just 1 entry

* fix order of commands

* remove newline

* measure cancelled extent read requests

* count all cancelled extent requests

* do not double count failed pages

* fixed cancelled name

* Fix error and warnings when compiling with --disable-dbengine

* when the timeframe is outside retention and whole query should fail

* do not mark as failed pages that have been loaded but have been skipped

* added chart to show cache memory calculation variables

* LONG_MAX for 32-bit compatibility

* fix cache size calculation on 32-bit

* fix cache size calculation on 32-bit - use unsinged long long

* fix compilation warnings on 32-bits

* fix another compilation warning on 32-bits

* fix compilation warnings on older 32-bit compilers

* fix compilation warnings on older 32-bit compilers - more of them

* disable ML threads joining

+++ database/engine/cache.c
@@ -310,7 +310,7 @@ static inline size_t cache_usage_per1000(PGC *cache, size_t *size_to_evict) {
 
     current_cache_size = __atomic_load_n(&cache->stats.size, __ATOMIC_RELAXED);
 
-    per1000 = current_cache_size * 1000 / wanted_cache_size;
+    per1000 = (size_t)((unsigned long long)current_cache_size * 1000UL / (unsigned long long)wanted_cache_size);
 
     __atomic_store_n(&cache->usage.per1000, per1000, __ATOMIC_RELAXED);
     __atomic_store_n(&cache->stats.wanted_cache_size, wanted_cache_size, __ATOMIC_RELAXED);
